Changed addfooditem view, when modifying existing selected item, amount is now shown italic in stead of grey because the grey didn't have enough contrast against the blue
Changed addfooditem view : details button (with amount of carbs calculated) now shows only the carbs 

Removed the text "chosen amount", and the amount, actually this is visible in the details button text. This makes the first change (actual chosen amount) useless.  Also when standard amount is 100, now it will be the chosenamount, so when clicking "add" without modifying the amount, then 100 will be chosen

":" added between "meal" and chosen meal when modifying an existing selected item (addfooditemview)

splash screen added