In all event screens, removed the save/add button, replaced by "done" icon top right corner Also added delete button, top right corner for existing events
In addfooditem view, also there the 'add' replaced by the "done" icon
In addfooditem view, when an existing item is being modified, then a trash button is added, when pressing this button, selectedfooditem is being deleted, and view is popped. If it's the last selectedfooditem, then also the mealevent is being deleted, and one more view is popped.

When selecting a unit with standard amount 100, and pressing the ok button without changing the amount, then 100 will be picked, always